requirements.txt
fastapi==0.88.0
pytest==7.1.3
pytest-asyncio==0.20.3
uvicorn==0.20.0
pydantic~=1.10.4
cfx-address
web3
python-dotenv
